Fix bugs

+++ uninstall.sh
@@ -5,6 +5,6 @@ if ! [ $HOME/.tirith ] ; then
 	exit 1;
 fi
 
-python3 ./aws/security_group.py delete < $HOME/.tirith/security_group.json
-
-rm $HOME/.tirith -r
+if python3 $HOME/.tirith/aws/security_group.py delete < $HOME/.tirith/security_group.json; then
+	rm $HOME/.tirith -r
+fi
